imagemagick-api/requirements.txt
# requirements.txt
flask==2.0.3
werkzeug==2.0.3
gunicorn==20.1.0
pillow==9.0.1
psd-tools==1.9.19
rawpy>=0.17.0  # 使用兼容ARM的最新版本
imageio==2.16.1